feat: bump packages and fix apps issues
