+++ apps/server/src/app.controller.spec.ts
@@ -3,19 +3,18 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 
 describe('AppController', () => {
-  let appController: AppController;
+  let app: TestingModule;
 
-  beforeEach(async () => {
-    const app: TestingModule = await Test.createTestingModule({
+  beforeAll(async () => {
+    app = await Test.createTestingModule({
       controllers: [AppController],
       providers: [AppService],
     }).compile();
-
-    appController = app.get<AppController>(AppController);
   });
 
-  describe('root', () => {
+  describe('getHello', () => {
     it('should return "Hello World!"', () => {
+      const appController = app.get(AppController);
       expect(appController.getHello()).toBe('Hello World!');
     });
   });
