+++ apps/server/tsconfig.json
@@ -1,5 +1,10 @@
 {
   "extends": "tsconfig/server.json",
-  "include": ["**/*.ts"],
-  "exclude": ["**/*.js", "**/*.json", "node_modules", ".turbo"]
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "./dist",
+    "baseUrl": "./src"
+  },
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+  "exclude": ["**/*.js", "**/*.json", "./test", "node_modules", ".turbo"]
 }
