+++ apps/web/.eslintrc.js
@@ -1,8 +1,4 @@
 module.exports = {
-  ...require('eslint-config-custom/eslint-next'),
-  parserOptions: {
-    root: true,
-    tsconfigRootDir: __dirname,
-    project: ['./tsconfig.json'],
-  },
+  root: true,
+    extends: ["custom"],
 };
