+++ apps/web/next.config.js
@@ -1,5 +1,4 @@
-const withTM = require("next-transpile-modules")(["ui"]);
-
-module.exports = withTM({
+module.exports = {
   reactStrictMode: true,
-});
+  transpilePackages: ['ui'],
+};
