+++ apps/web/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev -p 8080",
     "build": "next build",
     "start": "next start",
     "lint": "next lint"
@@ -15,13 +15,14 @@
     "ui": "workspace:*"
   },
   "devDependencies": {
-    "@types/node": "^18.0.0",
-    "@types/react": "18.2.0",
-    "autoprefixer": "^10.4.11",
+    "@types/node": "^18.11.17",
+    "@types/react": "18.0.26",
+    "@types/react-dom": "^18.0.9",
+    "autoprefixer": "^10.4.13",
     "eslint-config-custom": "workspace:*",
     "eslint": "8.39.0",
-    "next-transpile-modules": "10.0.0",
     "postcss": "^8.4.16",
+    "tailwind-config": "workspace:*",
     "tailwindcss": "^3.1.8",
     "tsconfig": "workspace:*",
     "typescript": "^5.0.0"
