--- apps/web/pages/_app.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import '../global.css';
-
-import type { AppProps } from 'next/app';
-
-function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
-}
-
-export default MyApp;
