--- apps/web/pages/index.tsx
@@ -1,10 +1,0 @@
-import { Button } from "ui";
-
-export default function Web() {
-  return (
-    <div>
-      <h1>Web</h1>
-      <Button />
-    </div>
-  );
-}
