+++ packages/eslint-config-custom-server/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "eslint-config-custom",
+  "name": "eslint-config-custom-server",
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
@@ -13,12 +13,9 @@
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-config-next": "^13.0.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-import-resolver-typescript": "^3.0.0",
     "eslint-plugin-import": "2.27.5",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "7.32.2",
-    "eslint-plugin-react-hooks": "^4.5.0"
+    "eslint-plugin-prettier": "^4.0.0"
   }
 }
