+++ packages/tailwind-config/tailwind.config.js
@@ -1,0 +1,12 @@
+/** @type {import('tailwindcss').Config} */
+module.exports = {
+  content: [
+    // app content
+    `src/**/*.{js,ts,jsx,tsx}`,
+  ],
+  theme: {
+    extend: {
+    },
+  },
+  plugins: [],
+};
