+++ packages/tsconfig/package.json
@@ -1,11 +1,9 @@
 {
   "name": "tsconfig",
-  "version": "0.0.0",
+  "version": "1.0.0",
   "private": true,
-  "main": "index.js",
-  "files": [
-    "base.json",
-    "nextjs.json",
-    "server.json"
-  ]
+  "license": "MIT",
+  "publishConfig": {
+    "access": "public"
+  }
 }
