--- packages/ui/index.tsx
@@ -1,2 +1,0 @@
-import * as React from "react";
-export * from "./Button";
