+++ packages/ui/package.json
@@ -1,19 +1,32 @@
 {
   "name": "ui",
-  "version": "0.0.0",
-  "main": "./index.tsx",
-  "types": "./index.tsx",
+  "version": "1.0.0",
+  "sideEffects": [
+    "**/*.css"
+  ],
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": "./dist",
+    "./styles.css": "./dist/index.css"
+  },
   "license": "MIT",
   "scripts": {
-    "lint": "eslint *.ts*"
+    "build": "tsup",
+    "dev": "tsup --watch",
+    "check-types": "tsc --noEmit"
+  },
+  "peerDependencies": {
+    "react": "^18.2.0"
   },
   "devDependencies": {
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
-    "eslint": "^8.0.0",
+    "@types/react": "^18.0.26",
+    "eslint": "^7.32.0",
     "eslint-config-custom": "workspace:*",
-    "react": "^18.0.0",
+    "postcss": "^8.4.20",
+    "react": "^18.2.0",
+    "tailwind-config": "workspace:*",
     "tsconfig": "workspace:*",
-    "typescript": "^5.0.0"
+    "tsup": "^6.1.3",
+    "typescript": "^4.9.4"
   }
 }
