+++ packages/ui/tailwind.config.js
@@ -1,7 +1,7 @@
-// tailwind config is required for editor support
 const sharedConfig = require("tailwind-config/tailwind.config.js");
 
-/** @type {import('tailwindcss').Config} */
 module.exports = {
+  // prefix ui lib classes to avoid conflicting with the app
+  prefix: "ui-",
   presets: [sharedConfig],
 };
